Resolve warnings 'sign-compare' which occurrs when comparing UINT to int. Pretty annoying because casting needs to be added. The logic I tried to follow: 1. In conditions prefer C-style casting to the longer static_cast 2. In conditions try to put cast on the right hand side of the equation and put it to minimize the amount of casting (which is why sometimes I cast to INT and something to UINT)
Worthy mentions:
1. Citizen.cpp - I also (hopefully) simplified the code a little bit by precalculating the limit

+++ DROD/GameScreen.cpp
@@ -4126,7 +4126,7 @@ SCREENTYPE CGameScreen::ProcessCueEventsBeforeRoomDraw(
 					wTileNo = this->pRoomWidget->GetEntityTile(eMonsterType, eLogicalType, pCoord->wO, 0);
 				}
 
-				if (pCoord->wValue2 != WT_Off) {
+				if (pCoord->wValue2 != (UINT)WT_Off) {
 					const UINT swordTile = this->pRoomWidget->GetSwordTileFor(pCoord->wValue - M_OFFSET, pCoord->wO, pCoord->wValue2);
 					if (swordTile) {
 						const int wSwordX = pCoord->wX + nGetOX(pCoord->wO);
